move grpc_server to mega_server
